feat: update backend api implementation
