+++ backend/api/core/dependencies.py
@@ -1,8 +1,8 @@
-from fastapi import HTTPException, Request, status
+from fastapi import HTTPException, status
 from fastapi.params import Depends
 from fastapi.security import HTTPAuthorizationCredentials, HTTPBearer
 
-from api.security.auth0 import VerifyToken
+from api.core.security import VerifyToken
 
 token_auth_scheme = HTTPBearer()
 
@@ -12,7 +12,3 @@ def verify_token(token: HTTPAuthorizationCredentials = Depends(token_auth_scheme
     if result.get("status"):
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail=result.get("msg", ""))
     return result
-
-
-def get_access_token(request: Request):
-    return request.headers["Authorization"].split()[1]
