+++ backend/api/models/user.py
@@ -1,9 +1,7 @@
-import logging
-
 from pynamodb.attributes import ListAttribute, UnicodeAttribute
 from pynamodb.models import Model
 
-from api.config import Settings, get_settings
+from api.core.config import Settings, get_settings
 
 settings: Settings = get_settings()
 
@@ -22,17 +20,3 @@ class UserModel(Model):
     name = UnicodeAttribute(null=False)
     room = UnicodeAttribute(null=False)
     reservations = ListAttribute()
-
-
-def create_table_if_not_exist():
-    """
-    Will creates the table if not exists
-    @return: True/False
-    """
-    if not UserModel.exists():
-        try:
-            UserModel.create_table(read_capacity_units=1, write_capacity_units=1, wait=True)
-        except Exception as e:
-            logging.error(str(e))
-            return False
-    return True
