+++ backend/api/routes/user.py
@@ -1,50 +1,48 @@
-from typing import List
-
 from fastapi import APIRouter, Depends
 
 import api.repos.reservation as reservation_repo
 import api.repos.user as user_repo
 from api.schemas.payload import Payload
 from api.schemas.user import User
-from api.security.dependencies import verify_token
+from api.core.dependencies import verify_token
 
 router = APIRouter(tags=["User"])
 
 
-@router.get("/user")
+@router.get("/user", response_model=list[User])
 def get_all_user():
     """Get all user objects from the database"""
     return user_repo.get_all_users()
 
 
-@router.post("/user")
+@router.post("/user", response_model=User)
 def post_user(user: User):
     """Create a user object to the database"""
     return user_repo.create_user(user)
 
 
-@router.get("/user/{user_id}")
+@router.get("/user/{user_id}", response_model=User)
 def get_user(user_id: str, token: Payload = Depends(verify_token)):
     """Get a user object from the database"""
     return user_repo.get_user_by_id(user_id)
 
 
-@router.patch("/user/{user_id}")
+@router.patch("/user/{user_id}", response_model=User)
 def update_user(user_id: str, user: User, token: Payload = Depends(verify_token)):
     """Update a user object from the database"""
     return user_repo.update_user(user_id, user)
 
 
-@router.get("/user/{user_id}/reservation")
+@router.get("/user/{user_id}/reservation", response_model=list[str])
 def get_reservation_date(user_id: str, token: Payload = Depends(verify_token)):
     """Get a user's reservation dates"""
     return reservation_repo.get_reservation_date(user_id)
 
 
-@router.put("/user/{user_id}/reservation")
+@router.patch("/user/{user_id}/reservation", response_model=list[str])
 def update_user_reservation(
     user_id: str,
-    reservation_date: List[str],
+    reservation_date: list[str],
     token: Payload = Depends(verify_token),
 ):
     """Update a user's reservation dates"""
